requirements.txt.txt
tensorflow==2.10.0
PyYAML==6.0.1
numpy==1.24.4
torch==2.2.1+cu118
torchvision==0.17.1+cu118
Pillow==10.2.0
matplotlib==3.8.3
opencv-python==4.9.0.80
scipy==1.12.0